app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.4.0",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
